% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-timeline-connector.R
\docType{data}
\name{geom_timeline_connector}
\alias{geom_timeline_connector}
\alias{GeomTimelineConnector}
\title{Timeline elbow connector geom}
\format{
An object of class \code{GeomTimelineConnector} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
geom_timeline_connector(
  mapping = NULL,
  data = NULL,
  stat = "timeline",
  position = "identity",
  elbowed = FALSE,
  connector_type = NULL,
  size = 0.4,
  colour = "grey50",
  linetype = 1,
  curvature = 0.3,
  show.legend = FALSE,
  inherit.aes = TRUE,
  ...
)

GeomTimelineConnector
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{elbowed}{If \code{TRUE}, draw horizontal-then-vertical elbows; if \code{FALSE}
(default), draw straight diagonal connectors. May also be \code{"curved"} as
an alias for \code{connector_type = "curved"}.}

\item{connector_type}{Connector style: \code{"straight"}, \code{"elbow"}, \code{"curved"},
or \code{"none"} (no connector drawn). When set, overrides \code{elbowed}.}

\item{size}{Line width.}

\item{colour}{Line colour.}

\item{linetype}{Line type.}

\item{curvature}{Curvature passed to \code{\link[grid:grid.curve]{grid::curveGrob()}} when
\code{connector_type = "curved"} (or \code{elbowed = "curved"}). Default \code{0.3}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}
}
\description{
Draws elbowed connector lines from axis points to label positions.
Requires data processed by \code{\link[=stat_timeline]{stat_timeline()}}.
}
\keyword{datasets}
