% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtimeline-package.R
\docType{package}
\name{ggtimeline-package}
\alias{ggtimeline-package}
\title{ggtimeline: Timeline charts for ggplot2}
\description{
Create publication-ready timeline charts with a thick arrow axis, boxed
labels, era bands, and automatic label placement.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brooksbenard/ggtimeline}
  \item Report bugs at \url{https://github.com/brooksbenard/ggtimeline/issues}
}

}
\author{
Brooks Benard (\email{brooksbenard10@gmail.com})
}
\keyword{internal}
