% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-timeline.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Timeline theme presets}
\usage{
theme_timeline(style = c("minimal", "nature", "dark"), ...)
}
\arguments{
\item{style}{Theme name: \code{"minimal"}, \code{"nature"}, or \code{"dark"}.}

\item{...}{Passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A ggplot2 theme object.
}
\description{
Timeline theme presets
}
\examples{
\dontrun{
ggtimeline(...) + theme_timeline("nature")
}
}
