% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtimeline.R
\name{ggtimeline}
\alias{ggtimeline}
\title{ggtimeline}
\usage{
ggtimeline(data, group, label, start, end, y_position, layout = "mirrored")
}
\arguments{
\item{data}{a data frame}

\item{group}{data to be displayed along a single track of the timeline}

\item{label}{text label for the starting point of each segment in the track}

\item{start}{starting position of a segment on the timeline}

\item{end}{end position of a segment on the timeline}

\item{y_position}{track position}
}
\value{
fdsa
}
\description{
ggtimeline
}
