% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtimeline.R
\name{ggtimeline}
\alias{ggtimeline}
\title{ggtimeline}
\usage{
ggtimeline(
  df,
  date_col,
  title_col,
  color_col = NULL,
  time_span,
  time_space,
  positions = c(0.5, -0.5, 1, -1, 1.25, -1.25, 1.5, -1.5)
)
}
\arguments{
\item{df}{Dataframe containing time series data}

\item{date_col}{Name of column containing dates}

\item{title_col}{Name of column containing timeline data}

\item{color_col}{Name of column that contains groups to color points}

\item{time_span}{One of "day", "month", or "year"}

\item{time_space}{Number of days/months/years between those shown on timeline}

\item{positions}{Vector of timeline point positions}
}
\value{
ggplot-based timeline
}
\description{
Creates timeline from time series data.
}
