% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtimeline.R
\name{timeline_scales}
\alias{timeline_scales}
\alias{scale_timeline_colour}
\alias{scale_timeline_fill}
\alias{scale_timeline_shape}
\alias{timeline_group_guide}
\title{Timeline scale helpers}
\usage{
scale_timeline_colour(palette = NULL, ...)

scale_timeline_fill(palette = NULL, ...)

scale_timeline_shape(...)

timeline_group_guide(
  guide = ggplot2::guide_legend(override.aes = list(linetype = 0, size = 4))
)
}
\arguments{
\item{palette}{Character vector of colours for timeline categories.
Defaults to \code{\link[=timeline_palette]{timeline_palette()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}
or \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.}

\item{guide}{A ggplot2 guide function, typically \code{\link[ggplot2:guide_legend]{ggplot2::guide_legend()}}.}
}
\description{
Timeline scale helpers
}
